backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
openai==1.3.0
speechrecognition==3.10.0
pyttsx3==2.90
python-dotenv==1.0.0
deep-translator==1.11.4
